put the Git helper code on the Git object
