+++ src/shared/Microsoft.Git.CredentialManager/CommandContext.cs
@@ -60,11 +60,6 @@ namespace Microsoft.Git.CredentialManager
         IGit Git { get; }
 
         /// <summary>
-        /// Helper process invoker.
-        /// </summary>
-        IHelperProcess HelperProcess { get; }
-
-        /// <summary>
         /// The current process environment.
         /// </summary>
         IEnvironment Environment { get; }
@@ -98,7 +93,6 @@ namespace Microsoft.Git.CredentialManager
                                             gitPath,
                                             FileSystem.GetCurrentDirectory()
                                         );
-                HelperProcess     = new HelperProcess(Trace);
                 Settings          = new Settings(Environment, Git);
                 CredentialStore   = new WindowsCredentialManager(Settings.CredentialNamespace);
             }
@@ -115,7 +109,6 @@ namespace Microsoft.Git.CredentialManager
                                             gitPath,
                                             FileSystem.GetCurrentDirectory()
                                         );
-                HelperProcess     = new HelperProcess(Trace);
                 Settings          = new Settings(Environment, Git);
                 CredentialStore   = new MacOSKeychain(Settings.CredentialNamespace);
             }
@@ -133,13 +126,12 @@ namespace Microsoft.Git.CredentialManager
                                             gitPath,
                                             FileSystem.GetCurrentDirectory()
                                         );
-                HelperProcess     = new HelperProcess(Trace);
                 Settings          = new Settings(Environment, Git);
                 IGpg gpg          = new Gpg(
                                             Environment.LocateExecutable("gpg"),
                                             SessionManager
                                         );
-                CredentialStore   = new LinuxCredentialStore(FileSystem, Settings, SessionManager, gpg, Environment, Git, HelperProcess);
+                CredentialStore   = new LinuxCredentialStore(FileSystem, Settings, SessionManager, gpg, Environment, Git);
             }
             else
             {
@@ -190,8 +182,6 @@ namespace Microsoft.Git.CredentialManager
         public IHttpClientFactory HttpClientFactory { get; }
 
         public IGit Git { get; }
-
-        public IHelperProcess HelperProcess { get; }
 
         public IEnvironment Environment { get; }
